Exercise 1.1
Create an XML

Create an XML file that stores data for a library. 
The library has two catalogues (technical books, and for cookery books).
Each catalogue can contain a number of books (say 2 for the purpose of this exercise) . 
Books will have an ISBN, title and author.
Upload you .xml file to here, and to a github repository called dataRepresentation
I will look for a link to this repository at the end of the course.